+++ src/types/nl2br.d.ts
@@ -1,5 +1,0 @@
-declare module "nl2br" {
-  function nl2br(str: string, isXhtml?: boolean): string;
-  namespace nl2br {}
-  export = nl2br;
-}